[libromdata] RomData: Initial RomData base class.
This class defines the types of fields and will be populated by
subclasses in order to allow a more generic UI to be written,
instead of having to write one UI per platform per ROM type.

MegaDrive currently fills in the field types and adds the string
fields, but not the numeric fields or bitfields.

TextFuncs: New functions for rp_char/rp_string:
- rp_strlen()
- rp_strdup() - returns an rp_char* duplicate.