[libromdata] ascii_to_rp_string(): Don't allocate len+1.
std::basic_string<> automatically allocates a NULL byte, so this
simply added a NULL byte to the actual string. Interestingly, this
only showed up on my ThinkPad T60p running KDE 4.x (not on KDE 5.x
in a VM), and only in an RFT_LISTDATA widget, where it showed a "NUL"
box at the end of the string.