[libromdata] Added query functions for extensions and supported image types; changed m_icon to m_images[].
This will allow for multiple types of images to be loaded.
Currently, NintendoDS is the only class that supports loading
images, and just IMG_INT_ICON.

NOTE: The query functions are virtual member functions instead of
base class non-virtual functions. This is how Qt implements it.
It has slightly more overhead, but it's much more flexible than
having the subclass initialize base class variables. I may rework
some of the existing functions to do this.

TODO: Rework isRomSupported() to be a virtual member function that
calls a static function so we can check with both a base class
pointer and hard-coded like in RomDataFactory?

[kde] RomThumbCreator:
- Updated for the image changes.
- Added "using namespace LibRomData;"