Initial implementation of IExtractIcon for Win32.
This implementation currently provides an IExtractIcon implementation
that returns a 32x32 cyan square. It provides a DllRegisterServer()
function that registers the CLSID of RP_ExtractIcon, the "rom-properties"
ProgID with IconHandler, and a ".nds" file type with the "rom-properties"
ProgID.

Lots of cleanup is still needed, and I don't have DllUnregisterServer()
implemented yet.