[win32] RP_ComBase: Added a virtual destructor.
gcc kept complaining about calling "delete this;" from a base class with
a non-virtual destructor. COM interfaces typically don't have a virtual
destructor, but this is an implementation of the interface, not the
interface itself.

I don't think this issue affected any of the COM classes currently
implemented, but we should fix it now anyway.

Other changes:
- RP_ClassFactory::LockServer(): Marked fLock as unused.
- RegKey: Reordered initializer list to match the declarations.
  I did this for the first constructor in commit 3088a40d,
  but forgot about the second.