[kde] RomThumbCreator: Initial external media fetching implementation.
This implementation attempts to fetch IMG_EXT_MEDIA if the RomData
subclass says it has it. I used cURL instead of QtNetwork because
QtNetwork only works asynchronously, which would be a pain to implement
with the ThumbCreator interface.

Currently, this downloads disc art for GameCube and Wii games from
GameTDB. It appears to work correctly, though parts of the disc
are white instead of transparent and hence I need to add some
post-processing.

*.desktop: Enabled thumbnail caching to reduce network traffic.

FIXME:
- Cache the GameTDB results ourselves in addition to the standard
  thumbnail cache. Otherwise, we'll end up downloading disc artwork
  multiple times if the same game is present in different locations
  on the system.
- Fix URLs for non-US region discs.
- Add automatic transparency cutting.