[libromdata] New IRpFile interface and RpFile class.
RpFile is an implementation using the system-native file I/O.
It's currently stdio, but will be changed to native Win32 API
for Windows.

Updated all libromdata classes to use RpFile instead of directly
using FILE*. This will make it easier to support other types of
files, including memory files and possibly KIO.

FIXME: RpFile_stdio.cpp does not currently compile in the Windows
build of libromdata8. Not a major issue, but it should be fixed.