[libromdata] NintendoDS: Fix icon extraction for the Sonic Colors demo.
Game ID NTRJ8P, sha1: 1dd6e7893ce6a82e001db2cbe7075e0b0300b7ef

The icon and title information are stored near the end of the ROM,
and parts of the title information are missing. (Only 2,248 bytes
were present; 0x8C8, which is in the middle of title 6.)

Changed the check to ensure that we have the icon data and at least
two titles (Japanese and English).

DS_IconTitleData: Fixed some fields and added a static_assert() for
the struct size in loadInternalImage().