[libcachemgr] Initial CacheManager implementation.
CacheManager provides an interface to specify a URL and a cache key.
It then checks the local cache for a matching key. If an entry is found,
it either returns the filename, or an empty string if the cached file
is 0 bytes, which indicates the file didn't exist the last time a
download was attempted.

[libromdata]
- Added _RP_CHR() for character constants, which is required due to
  older MSVC (pre-2015) not supporting Unicode character constants
  from C++ 2011.
- IRpFile: Added a write() function. This is used when caching the
  image files.

FIXME:
- Need to implement various functions in CacheManager for Windows.
- Add a "fallback" cache entry for GameTDB. GameTDB usually considers
  both GCN and Wii as "wii", but we're using separate directories. If
  a file is not found in the GCN cache directory but is found in Wii,
  then the Wii file should be used.
- Don't negatively cache results other than 404.