[libromdata] RomData: Added a fallback cache key.
This is useful for GameTDB, which groups GCN, Wii, and Triforce games
all as "wii". If a GameTDB pack is downloaded and everything is saved
in the "wii" directory, those images will be used.

If the offline pack isn't downloaded, the GCN files will be saved to
the "gcn" directory, which takes precedence over GameCube images in
the "wii" directory.