[win32] RP_ShellPropSheetExt: Initial rework using WM_INITDIALOG control creation.
Dialog units have rounding issues, resulting in uneven grids. I noticed
this when attempting to implement bitfield checkboxes using DialogBuilder.
With other controls, it's less noticeable, but it's still there.

This rewrite still uses dialog units as the base unit in order to work
on high DPI systems, but it uses it as a static grid, so spacing between
rows and columns are always the same.

Compared to the previous version, some rows are spaced an extra pixel
apart, and the description column is one pixel narrower. (Tested on
Windows XP using the default Luna theme at 96 dpi.)

TODO: Rework RFT_BITFIELD and RFT_LISTDATA to handle these changes.