[libromdata] MegaDrive: Rewrote to use a private class.
Moved the RomFields definitions into the private class.

Moved MD_RomHeader into the private class and marked it as PACKED.
(The struct elements were aligned before, so marking it as PACKED
won't actually affect it, though it's good practice to ensure that
the struct won't be "optimized" by the compiler.)

Split I/O Support and Region Codes parsing into their own functions
in the private class.