[libromdata/tests] RpImageLoaderTest: Added BMP reference images.
The BMP images are gzipped to reduce their size, though they're still
quite big compared to the PNG images, which is amusing because PNG
uses zlib for its own compression.

Currently, the BMP images are loaded into memory, and the fields in
BITMAPFILEHEADER and BITMAPINFOHEADER are verified against known values
in the test parameters.

TODO: Compare the image data in the decoded PNG image (rp_image) and the
reference BMP image.