[cmake] msvc.cmake: Disable /NOLOGO for RC and MASM; use uppercase switches.
Only CL explicitly requires /nologo to be lowercase. The rest of them
are supported better by CMake if they're uppercase.

CMake's MSBUILD generator was already adding /NOLOGO to RC, so adding
it again caused RC to go crazy and think that we specified "-ologo".

Also disabled /NOLOGO for MASM, just in case.