[libromdata] RpPng_gdiplus.cpp: Handle grayscale images correctly.
GDI+ converts grayscale PNG images to ARGB32 for some reason.
It does set a flag indicating that the image is graysclae, so if we get
an ARGB32 image from GDI+ and it has the ImageFlagsColorSpaceGRAY flag
set, convert the image:

- Set the rp_image's format to FORMAT_CI8.
- Keep the Gdiplus::Bitmap as PixelFormat32bppARGB, since specifying
  PixelFormat8bppIndexed results in a Status::InvalidArgument error.
- When copying the image, create a default 256-color grayscale palette,
  then copy the least-significant byte (blue) from each ARGB32 color
  and use that as the grayscale value.

This fixes the two "gray" tests in RpImageLoaderTest on Windows.