[libcachemgr] IDownloader: Added an mtime() field.
CurlDownloader: Set m_mtime to the "Last-Modified:" header
value
if it's present. This uses curl_getdate() to parse the header
value into a standard time_t.

CacheManager: If the downloader's mtime is >= 0, set the
cached
file's modification time to the mtime. This indicates that the
downloader retrieved and parsed a "Last-Modified:" header.

NOTE: curl_getdate() handles the correct format without interference
from the locale. strptime() uses locale-dependent strings, so it
won't work for HTTP headers.

The KDE4/KDE5 build now sets the mtime on downloaded files to the
value provided by the server.

TODO: Implement this using WinInet on Windows.