[libromdata/tests] Moved RpImageLoaderTest to the img/ subdirectory.
The various image files were starting to clutter the tests/ directory.

Note that the images are still copied to the tests/ directory in the
actual build, so the build directory will still be cluttered.

The PNG tests are now in RpPngFormatTest.cpp, leaving only the main()
function in RpImageLoaderTest.cpp.