[libromdata/tests] RpPngFormatTest: Added an "alternate" format.
For some reason, GDI+ on wine-1.9.18 (and previously 1.9.16) and AppVeyor
is converting gl_triangle.gray.png and gl_quad.gray.png to ARGB32. Well,
it always does that, but it's setting the alpha flag in the image flags,
which prevents us from converting the image to grayscale. We'd have to
scan the entire image for alpha values, which ends up taking longer than
simply processing the ARGB32 image as-is.

Added a comparison function Compare_ARGB32_BMP8() to compare ARGB32
rp_image objects to CI8 bitmaps. This is now used in this situation.

Added a wrapper around gtest's internal ColoredPrintf(), which is used
to print a warning if the alternate rp_image format is being used.
Reference: http://stackoverflow.com/questions/16491675/how-to-send-custom-message-in-google-c-testing-framework