[libromdata/tests] Added CI8+tRNS image tests.
xterm-256color.CI8.tRNS is the same as xterm-256color.CI8, but with
a transparent background.

NOTE: GDI+ decodes a 256-color PNG with tRNS chunk as if it's an
ARGB32 image, so the rp_image will be ARGB32, not CI8.
- ...except on wine-1.9.18, which loads the image as CI8 with the
  correct transparency values. An alternate format is used as well
  as a new comparison function, Compare_CI8_BMP32().

No changes needed for either of the libpng or GDI+ PNG decoders,
though a note was added to RpPng_gdiplus.cpp regarding how it
decodes these images.

Added test images for both libpng (256-color bitmap) and GDI+
(32-bit color bitmap). The libpng version also has a tRNS chunk
defined within the RpPngFormatTest.cpp file, since BMP doesn't
support alpha values within the palette. (The colors are 32-bit,
but the extra byte is reserved and is always 0.)