[libromdata/tests] RpPngFormatTest: Added a 16-color paletted PNG with an odd width.
256-color BMPs must have an internal width that's a multiple of 8 pixels,
so we need to skip over those pixels when comparing the image.

RpPng_libpng.cpp: Convert 1, 2, and 4bpp paletted images to 8bpp.

RpPng_gdiplus.cpp: Convert 4bpp paletted images to 8bpp manually, since
GDI+ on Windows XP doesn't want to do that for us. (NOTE: On Wine, GDI+
loads the image as if it's ARGB32, so no conversion is needed.)