[libromdata] KeyManager: Initial encryption key manager class.
This class will be used to load encryption keys from the keys.conf file
in the user's configuration directory.

options.cmake: Added an option ENABLE_DECRYPTION. This is enabled by
default; if disabled, KeyManager won't be built, and decryption won't
be available for newer ROM and disc images that are either fully
encrypted or have encrypted sections.