[libromdata] MegaDrive: Added support for SMD-format ROMs.
The rest of the infrastructure was previously implemented, including
file extension registration on Windows, but the actual deinterleaving
was not.

NOTE: Only SMD-format MD ROMs are supported. SMD-format 32X ROMs are not
supported. Among other things, you can't actually *use* a Super Magic
Drive copier with a 32X connected to a Mega Drive.

New function decodeSMDBlock(), which decodes a 16 KB SMD-format block.
Note that it does not support in-place decoding. (TODO: Add the "restrict"
keyword to both parameters?)

MegaDrive constructor: Deinterleave the first block of SMD-format ROMs
and load the MD ROM header.