[libromdata] Added an explicit RomFields::addData_invalid() function.
Previously, addData_string(nullptr) was used to hide a field. This was
only used for non-RFT_STRING fields, so it worked because the data type
didn't match the field description. It didn't work for hiding string
fields. Now, RFT_INVALID is used to explicitly hide the field.

Updated all RomData subclasses to use addData_invalid() where necessary.

Updated kde/ and win32/ to handle this change.