[libromdata] WiiPartition: Reference-count aes_common.
aes_common[] is now an array of two AesCipher objects. Index 0 is the
Wii common key; index 1 is the Korean key.

A single reference counter is used for all WiiPartition objects.

Split the decryption code into a new function, initDecryption().
The partition header is read in the constructor. For now, the
initDecryption() function is called in the constructor, but later
it will only be called in read(), which will allow the partition
header to be read without initializing the keys.

TODO: Atomic reference count?