[libromdata] Converted IDiscReader into a pure virtual interface.
This moves the IRpFile* pointer into subclasses, which will allow us to
create a new IPartition interface for WiiPartition that derives from
IDiscReader.

NOTE: A definition for IDiscReader::~IDiscReader() is still needed, even
though it's pure virtual. Both gcc and MSVC complain about undefined
references if it's not there, but they're perfectly fine with it being
declared as inline within the IDiscReader.hpp header.