[libromdata] GcnFst: Preliminary GameCube/Wii file system table parser.
WiiPartition: Added a GcnFst object, but it isn't usable until the
read() function is implemented. Also, we shouldn't load the FST
unless it's actually needed, so we should add a function to
IPartition for initializing decryption and/or file access.

gcn_structs.h: New structs GCN_FST_Info and GCN_FST_Entry.