[libromdata] SystemRegion::getCountryCode(): New function to get the system's country code.
On Windows, this uses GetLocaleInfo() with LOCALE_SISO3166CTRYNAME.

On Unix/LInux, this calls setlocale(), then looks for the country code
portion of the locale (e.g. the "US" in "en_US").

If the country code could not be determined, this function returns 0.

NOTE: SystemRegion is a static class. More functions might be added later.