[libromdata] MegaDriveRegions::parseRegionCodes(): &&, not &.
On Linux (glibc), isalnum() returned 8 for 'A' (PAL Chaotix), which
caused the bitwise comparison to fail, since the following expression
returns either 0 or 1.

On Windows (MSVC 2010), isalnum() returned 1 for 'A', so it just happened
to work correctly.

Using && instead of & fixes parseRegionCodes() on Linux when parsing
MD hexadecimal region codes.