Enable RTTI in the MSVC build.
We're going to use dynamic_cast<> for rp_image_backend, so RTTI needs
to be enabled. Otherwise, things will go haywire at runtime.

Also, removed the commented-out RTTI lines from gcc.cmake, since we're
not going to be reenabling it. (It was disabled for gcc builds because
KDE's thumbnail.so uses it for ThumbCreator subclasses.)