[libromdata] rp_image_backend: Changed data and data_len to pure virtual functions.
QImage is implicitly shared, so if there's two copies of a QImage and one
is modified, it has to be detached. This can change the data pointer.

We can't easily determine when the data pointer is changed, so let's use
a virtual function to retrieve it instead. This ensures that we always
have the correct data pointer.

data_len was changed to a virtual function for consistency, since it's
possible that the length changes when the pointer changes.