[libromdata] CisoGcnReader: IDiscReader class for GCN/Wii CISO files.
This is a sparse image format supported by Dolphin and a few USB loaders.
Not nearly as common as WBFS, though.

This CISO is also unrelated to the PSP CISO format, which has zlib
compression in addition to sparse.

NOTE: isRomSupported() can't determine the system type, since the disc
header is at $8000, so it sets the system type to unknown while setting
format to CISO. In this case, the GameCube constructor will check the
magic values itself to determine the actual system.

Added file extensions ".ciso" and ".cso".

TODO:
- Write an IDiscReaderFactory class to auto-detect the different types
  of disc image formats and return an IDiscReader.
- Add support for Dolphin GCZ images. (requires zlib)
- Maybe add support for PSP CISO images later, if I ever implement
  support for PSP disc images.