[libromdata] GameCube: ESRB ratings with 0x20 set indicate online play.
This might also apply to other ratings, but at least with Super Smash
Bros. Brawl, the US version has it set, but the PAL version does not.

Mask the rating with 0x1F before displaying it. This fixes the display
of Brawl (USA)'s rating, which now shows 13 (0x0D) instead of 45 (0x2D).

If 0x20 is set, display a "°" symbol after the rating.
- TODO: Explain this flag, and/or pick a better symbol.