[libromdata] GameCube: Use the Game ID region as a sub-region for the BI2 region.
The BI2 region is primary, but if e.g. the BI2 region is PAL and the
Game ID region is 'D' (Germany), it will now show "Germany (PAL)".

This also helps for distinguishing between regular Japanese games and
games released in Taiwan, which have a 'W' game ID region.

The BI2 region code is displayed parentheses if it doesn't match
the Game ID region.