[libromdata] GameCube: Split the GameTDB region code function into a separate function.
gcnRegionToGameTDB() is similar to gcnRegionToString() in that it checks
both the BI2 region and the Game ID region, so it should be more accurate
than just checking the Game ID region.

TODO: Add a user-specified fallback region.

gcnRegionToString(): Added 'C' for China. Not sure if it's actually used
by any commercial game or if it's just for unofficial translations.