[libromdata/tests] GcnFstTest.cpp: Initial GcnFst test.
No actual tests yet, but it does load the list of files from
GameCube.fst.zip and attempts to run tests.

FIXME: GMBE8P (Super Monkey Ball) seems to have a broken directory
structure as parsed by GcnFst. The first test will be to check if
any duplicate filenames are present in any subdirectories, which
indicates that something is wrong.