[libromdata/tests] GcnFstTest: New function getFileFromZip().
This is basically the file loading code from SetUp(). One major change
is that it uses EXPECT_*() instead of ASSERT_*(), which allows it to
close the Zip file on error. Hence, we don't need m_unzFstZip anymore.

The split is in preparation for adding an FstPrint test, which needs
to load a known-good printout from a second Zip file.