[libromdata/tests] Fixed some leak issues in FstPrint and GcnFstPrint.
GcnFstPrint: Free memory and delete the GcnFst when finished.
Not strictly necessary, but this fixes some warnings in cppcheck-1.75.

FstPrint: Close the IFst::Dir when finished printing the directory.
This is an *actual* memory leak, and it causes an assertion when the
IFst is deleted because the directory reference count isn't 0.