[win32] RpImageWin32::toHBITMAP_mask(): There's only an AND mask here.
The XOR mask is only present if the icon being masked is monochrome.

Note that we were clearing the XOR area using memset(), even though
CreateDIBSection() didn't allocate this. I'm guessing it "worked"
because of a quirk for this behavior for 32x32 icons. When attempting to
rework RP_ExtractIcon to use the new image extraction functions, Explorer
crashed when it encountered a Wii game disc image because it attempted to
create a 160x160 mask, and the memset() went out of bounds.