[win32] RpImageWin32::toHBITMAP_mask(): Fixed mask generation for ARGB32 images.
The mask bit should be 1 for opaque pixels and 0 for transparent.

This fixes an issue where on Windows XP, the icon for GCN and Wii discs
showed up fine in Windows Explorer, but was screwy in file properties.
(I didn't test this on Windows 7.)

TODO: Proper alpha transparency for ARGB32 icons.