[libromdata] TextFuncs: Support NULL-terminated strings; remove trailing NULLs.
The 'len' parameter can be set to -1 to indicate a NULL-terminated
string is specified, in which case strlen() will be used to determine
the actual string length.

If 'len' is specified as an actual string length (>0), it will be
used as an explicit string length, but trailing NULLs will be
discarded. This helps for some fields that use NULLs as a padding
character.

TODO: Space elimination for e.g. MD.