RP_ShellPropSheetExt: Work around ES_MULTILINE behavior for VK_RETURN and VK_ESCAPE.
ES_MULTILINE edit controls are now subclassed to override the default
behavior. VK_RETURN sends a WM_COMMAND+IDOK to the property sheet dialog,
while VK_ESCAPE sends a WM_COMMAND+IDCANCEL.

This is what normal edit controls do, but ES_MULTILINE overrides it.
(VK_RETURN is supposed to only be overridden if ES_WANTRETURN is set,
but Windows 7 doesn't seem to care.)

The property sheet dialog's hWnd is the parent of the property sheet.
This is now saved in WM_INITDIALOG as d->hDlgProps, and is used in the
MultilineEditProc subclass procedure to send the window messages.

Reference: http://blogs.msdn.com/b/oldnewthing/archive/2007/08/20/4470527.aspx