[libromdata] ImageDecoder: Set the transparent color for CI8 images.
The black squares seem to have disappeared on Windows 7 even without
this commit, but this commit helps to create a better icon mask.

TODO: Check the palette for alpha values other than 0 and 255,
and if found, convert to ARGB32 before rendering, since Windows
doesn't like CI8 images with alpha transparency.