[libromdata] TextFuncs.hpp: Handle len < 0 in the inline functions.
The MSVC build crashed when attempting to display a WBFS image because
changed some of the TextFuncs calls to use -1 instead of an explicit
strlen() or wcslen() call. The inline functions didn't handle this,
so they ended up passing -1 directly to the std::string or std::u16string
constructor, resulting in failure.

Changed some usage of std::string and std::u16string in TextFuncs.hpp to
rp_string. This will help to find type mismatches.

(It sure would be nice if I could commit a cleanup patch without having
it massively break the build...)