[libromdata] GcnPartition: New function open() for opening a file within the partition.
This function returns a PartitionFile, which is an IRpFile implementation
that uses an IPartition as its backend.

IFst::find_file(): New public function to find a file using the internal
find_path() function. The entry is returned using a DirEnt.

GcnPartition::open() uses GcnFst::find_file() to find the file. The
DirEnt information is then used to create a PartitionFile.

TODO: Reference count the PartitionFile and assert() if the underlying
IPartition is deleted before all PartitionFiles are closed.