[libromdata] RomDataView: Stop the animation timer when the tab isn't visible.
There's no point in updating the animation when the tab isn't visible.
This should reduce CPU usage a bit.

Added a new variable anim_running that's set to false to make sure
the timer stops. This should help in cases where the timer function
is running when stopAnimTimer() is called, though I don't think
that's likely.