[libromdata] NintendoDSPrivate::loadIcon(): Don't unconditionally return frames[0] as the default icon.
For Zelda Four Swords, the first icon in the sequence is actually
icon #5, which has all four swords onscreen. Icon #0 only has one
sword all the way on the right.

The correct first icon from the sequence is now copied and returned.
This copy is owned by the RomData superclass.

NintendoDS now owns all of the icons in the IconAnimData struct.

TODO:
- rp_image assignment operator and copy constructor.
- Configuration option to use the standard NDS icon instead of the
  first frame of the animated DSi icon. (Except for DSiWare...)