[libromdata] Split IconAnimData into its own file; added a new IconAnimHelper class.
IconAnimHelper handles the icon sequence and frame numbers.
The caller merely has to set the correct delays and update the
icon when the frame number changes.

IconAnimHelper::frame() returns the last valid frame, so the caller
doesn't have to worry about empty frames, either.

This will make it a bit easier to add animated icon support to the
Windows frontend.

[kde] RomDataView: Use IconAnimHelper.