[libromdata] RpGdiplusBackend: New functions dup(), lock(), and unlock().
dup() function clones the GDI+ bitmap and returns it.

Since this function is intended to be used when drawing GDI+ bitmaps
directly to a window, it will automatically convert CI8 images to
ARGB32 in order to avoid CI8 alpha transparency artifacting.

lock() and unlock() call m_pGdipBmp->LockBits() and m_pGdipBmp->UnlockBits(),
respectively. This helps to reduce code duplication.