[win32] RP_ShellPropSheetExt: Implemented animated icons.
Using SetTimer() with the property sheet hWnd for timing.

Most of the code was copied from the KDE RomDataView, with changes
for GDI+ and SetTimer(). Among other things, I'm using an array of
unique_ptr<Gdiplus::Bitmap*> instead of a plain array.

FIXME: The icon flickers occasionally, especially if the window is
being moved around.