[libromdata] KeyManager: Reload keys.conf if the file is changed.
This allows the user to manipulate keys.conf and have the changes take
effect when opening a new file properties dialog instead of having to
close the entire file browser (and/or kill the process).

NOTE: If the file was previously loaded and then deleted, the keys
will still remain in memory. They'll only be cleared and reloaded
if a new file is created.

FileSystem: New function get_mtime(). This function gets a file's
modification time.