[win32] Split COM registration into two functions: RegisterCLSID() and RegisterFileType().
This is needed in order to register the shell extension as a handler
for programs that the user already associated files with.

RegisterFileType() takes an optional parameter that indicates which
ProgID should be written to. If nullptr, RP_ProgID[] will be used.

RP_ExtractIcon::RegisterFileType(): Return lResult if the "DefaultIcon"
write fails, not SELFREG_E_CLASS.