Added git version checking.
CMakeLists.txt: Run git_version.sh. This requires a POSIX shell and
git, and both must be present in PATH. If the script can't be run,
a dummy git.h will be generated in the top-level binary directory.

libromdata, win32: Added the top-level binary directory to the
private include directory list.

[win32] resource.rc: #include "libromdata/git.h"
Removed the version string from FileDescription, since it's redundant,
and it gets cut off on Windows 7.

TODO: The git version gets cut off on Windows 7, since it uses a
ListView instead of separate ListBox and Edit controls that were
used on Windows XP.