[win32] Register file type handlers per extension instead of using a ProgID.
ProgID implies we have a program that opens these files, but we don't
actually support that. By registering as file type handlers instead of
a ProgID, we allow other programs to take over as the default program
while maintaining the extra shell functionality.

DllRegisterServer() will delete the 'rom-properties' ProgID if it's
still present, since it's no longer used.

TODO: Remove the file type handlers in DllUnregisterServer.