[scripts] package.cmd: Check for MSVC 2010, 2012, 2013, and 2015.
Highest version is preferred.

For MSVC 2012+, set the toolchain to e.g. v120_xp for XP compatibility.

Check for cmake.exe. CMake 3.0.0 is now required due to a naming change
for the MSVC generators. (The old names for 2010, 2012, and 2013 are
still supported, but I'd rather not use them.)

NOTE: We aren't checking the CMake version. Older versions will fail
due to an unsupported generator name.