[win32] RP_ExtractIcon: Always specify GIL_DONTCACHE.
If we don't set GIL_DONTCACHE, the first icon extracted will be used
as the icon for all handled files. I have no idea why.

The issue was showing up in Release builds in both VC10 and VC14,
but not debug builds, which makes sense because GIL_DONTCACHE was
in an NDEBUG conditional.

This fixes #27. (IExtractIcon only uses icon from the first file in VC14 builds.)