[win32] RP_ShellPropSheetExt: Re-create the monospaced font if the ClearType setting changes.
See issue #28. (ROM Properties tab uses mix of bitmapped and aliased fonts on Windows)

NOTE: We can't handle WM_SETTINGCHANGE because the property sheet isn't a
top-level window. WM_NCPAINT seems to work fine on WinXP and Win7. It's
sent on window creation and whenever the ClearType setting is changed.