[libromdata] RomData: New struct HeaderInfo.
This is now used in DetectInfo. It makes it possible to indicate that
the header data doesn't start at 0.

Updated all RomData subclasses for this change. Note that the subclasses
will all assert if headerInfo.addr != 0, since the ones currently
implemented all expect the header to be at 0. This will mainly be
used for SNES and other formats that use "footers" instead of "headers".

NOTE: Not asserting on headerInfo.size, since some files might be too
small for a given RomData subclass while being acceptable for another.

Other changes:
- Don't set DetectInfo.szFile if it isn't used by isRomSupported().
- Mark some ROM header pointers as const *const.
- Rename 'header' to 'romHeader' in some cases where a struct is used
  instead of a generic uint8_t array.