[libromdata] IRpFile: New virtual function filename().
The default implementation returns an empty string.
TODO: Return a const string& once this is changed to pure virtual.

RpFile: Save the filename and return it in filename().

RomDataFactory: Set info.ext to the file extension. This will be used
for some types that don't have magic numbers, but do have well-known
file extensions.