[libromdata] openDreamcastVMSandVMI(): delete test_file, not other_file.
other_file points to vmi_file or vms_file, depending on which one was
opened. It isn't set until the correct "other" file was opened, and
that file is stored in test_file.

This only affected the Linux build, since we check for both uppercase
and lowercase extensions due to case-sensitive file systems.