[win32] DllMain.cpp: Condensed some code.
DllGetClassObject(): Use a macro to generate the code for the various
COM interface implementations.

Other functions: Condense the SELFREG_E_CLASS checks into a single line.
Fixed some error paths to return SELFREG_E_CLASS. This is the only value
that should be returned by the exported functions on error.