[libromdata] ImageDecoder: New function to blit a CI4 tile directly to a CI8 image.
This should improve performance a bit because we no longer have to
convert the tile from CI4 to CI8 before blitting.

NOTE: BlitTile_CI4_LeftLSN() assumes the least-significant nybble is
the left pixel. This is true for Nintendo DS, but might not be true
for other tiled CI4 images, in which case a LeftMSN / RightLSN variant
will need to be added.

fromNDS_CI4(): Use BlitTile_CI4_LeftLSN().