[libpng] Initial import of libpng-1.6.23 from GCN MemCard Recover.
I've decided to get rid of the GDI+ PNG handling. It doesn't improve
anything performance-wise, adds coding overhead, and probably isn't
any faster than libpng.

GDI+ also doesn't support APNG. This version does.

NOTE: If I add APNG export, I'll need to use the included libpng
in Ubuntu builds, since Ubuntu still has libpng-1.2, and it's a
version that doesn't support APNG.

Reworked the bundled library system to be similar to the one in
mcrecover, but improved.

Added "-fpic -fPIC" to all extlib static libraries when building
for Linux systems. This is needed for libpng, but might not be
needed for other libraries; nevertheless, it should be set.

TODO: Use libpng in the Windows build. It's being linked in,
but it isn't actually being used for anything yet.