RpMemFile: Fixed issues reading out of bounds.
RpImageLoader reads the first 256 bytes of the image to determine what
kind of file it is. odd-width.16color.CI4.png and happy-mac.mono.odd-size.png
are both less than 256 bytes. However, the boundary check for RpMemFile
was using unsigned arithmetic, which prevented it from detecting an
out-of-bounds read in this case.

Cast the size_t values to int64_t to fix this issue.

This caused an SEH exception when running RpImageFormatTest on Windows XP
when using libpng instead of GDI+:

[ RUN      ] happy_mac_mono_png/RpPngFormatTest.loadTest/happy_mac_mono_odd_size_png
unknown file: error: SEH exception with code 0xc0,000,005 thrown in the test body.
[  FAILED  ] happy_mac_mono_png/RpPngFormatTest.loadTest/happy_mac_mono_odd_size_png, where GetParam() = happy-mac.mono.odd-size.png (10 ms)