[libromdata/tests] CMakeLists.txt: Copy test images to ${CMAKE_CURRENT_BINARY_DIR} in addition to bin/tests/.
CTest uses ${CMAKE_CURRENT_BINARY_DIR} as the working directory,
even if the executable is in bin/tests/, so we have to copy the
images there, too.

This fixes RpImageLoaderTest when run through CTest (and travis-ci and
AppVeyor).

This fixes a regression from commit 02315a1c1be07a2d840c4fffbc8a4595fb4419b6.
(Compile zlib and libpng as DLLs on Windows.)