[libcachemgr] CacheManager::downloadAndScrape(): New function to download and scrape image URLs.
This is used for IMGPF_EXTURL_NEEDS_HTML_SCRAPING, which is used by the
Amiibo class because http://amiibo.life/ doesn't provide direct URLs
for the image files. The RomData object needs to be passed to this
function in order to scrape the image URL correctly.

Missing images are now negatively-cached correctly, and the HTML pages
are no longer cached. The HTML pages are no longer downloaded and/or
received from the cache if the image files are present.

We don't need to add a function to check if the file exists in the cache,
since all of the cache handling is done in libcachemgr instead of in the
UI frontend.

[kde] RomThumbCreator: Call cache.downloadAndScrape(). All of the URL
scraping code has been moved to CacheManager.

TODO: Function pointer for image URL scraping instead of a RomData object?