Mostly fixed amiibo support on Windows.
TextFuncs: New implementation of memmem() for systems that don't have it.
This function is used by Amiibo::scrapeImageURL().

CMakeLists.txt, config.libromdata.h.in: Check for memmem() and define
HAVE_MEMMEM if it's present.

RpImageWin32::getExternalImage(): Call cache.downloadAndScrape() if the
image needs URL scraping.

This mostly fixes amiibo thumbnails on Windows, except for an assertion
in RpImageWin32::toHBITMAP_mask() that occurs for any thumbnail that
isn't a multiple of 8px in width. (This never happened before because
all previous thumbnails were multiples of 8px wide.)