[libromdata] pngcheck.cpp: Un-static some variables.
There's no reason why these variables should be static, especially
since pngcheck() is supposed to be reentrant.

This caused a crash due to a race condition when testing lots of amiibo
NFC dumps in a single folder on Windows 10, probably because Windows 10
(or Windows 8?) is using multiple threads in a single COM Surrogate,
whereas Windows 7 doesn't.

This bug was reported by @cheatfreak47 in #retro.