[zlib] CMakeLists.txt: Fix installation of PDB files in MSVC builds.
CPack doesn't like using the PDB property for some reason, so we have to
manually build the pathname using the correct postfix. Also, install both
debug and release PDBs with the "optional" tag so it won't fail if one is
missing (and one is always missing because we're only building either
Debug or Release).