Use static_cast<> instead of reinterpret_cast<> where possible.
Mostly when converting from void* to a different pointer or vice-versa
(e.g. for malloc()), and also when converting between primitive types.

static_cast<> is the preferred cast. reinterpret_cast<> should only be
used when converting between pointers that the compiler considers to be
"different", e.g. wchar_t* and char16_t* on Windows.

Reference: http://stackoverflow.com/questions/332030/when-should-static-cast-dynamic-cast-const-cast-and-reinterpret-cast-be-used