[libromdata] GameCube: Display the encryption key in use for Wii partitions.
WiiPartitionPrivate::getEncKey(): New function that determines the
encryption key based on the ticket. This returns a value of type
WiiPartition::EncKey, which maps directly to key indexes (or -1 if
the key is unknown).

WiiPartitionPrivate::initDecryption():
- Use getEncKey() to get the encryption key.
- Get the "rvl-korean" key if the key index is Korean. Previously,
  "rvl-common" was always used, even if the key index is 1.
- Removed some unused code that obtained the Wii common key even
  after it was already obtained.

WiiPartition::encKey(): New function that gets the WiiPartition::EncKey
value. This is used by the GameCube class to determine the partitions'
encryption keys.

NOTE: WiiPartitionPrivate::getEncKey() is needed because initDecryption()
is only called if actual partition data is requested. We want to be able
to get the encryption key of all partitions, even if we're not reading
data from the partition. (We're only reading from the Update partition
right now in order to determine the System Update version.)