[libcachemgr] CacheManager: Filter out invalid characters in the cache key and prevent directory traversal.
Characters filtered out include all control characters and reserved
characters on Windows, except for '.' (needed for file extensions)
and '/' (needed for directory hierarchies).

To prevent directory traversal attacks, we also check for "../". We
also check for the key ending in "..", since that could be a problem.
There's usually no reason for ".." to be at the end of the cache key,
since there's usually a file extension.

This should prevent weird issues on Windows if a screwy GameCube disc
ID is encountered, plus it will prevent directory traversal exploits.

I don't think there are any traversal exploits right now, since the
only use of the cache key is the 6-character GCN/Wii disc ID, and
that's prefixed by a GameTDB directory path, so there's not much
that can be traversed here.

Example: wii/disc/US/../../ would result in wii/, which can't do
much other than maybe resulting in a broken cache.